+++ server/public/styles.css
@@ -15,7 +15,7 @@ body {
 
 .player {
   background-color: #2d2424;
-  max-height: 90vh;
+  max-height: 80vh;
   margin-bottom: 0.5rem;
 }
 
@@ -46,4 +46,33 @@ body {
 
 .rounded {
   border-radius: 0.5rem;
+}
+
+.avatar {
+  margin: 0.25rem;
+  border-radius: 5rem;
+  background-color: #0d7d77;
+  width: 5rem;
+  height: 5rem;
+}
+
+.avatar > img {
+  width: 100%;
+  height: 100%;
+  object-fit: fill;
+  border-radius: 5rem;
+}
+
+.border-red {
+  border: 3px solid rgb(255, 0, 0);
+  box-shadow: 0 0 4px 2px  rgba(0, 0, 0, 0.226);
+}
+
+.profile {
+  display: flex;
+  font-family: 'Ramabhadra', sans-serif;
+}
+
+.profile span {
+  margin: 1rem;
 }
